feat: expose the OptionalVec helper object
